Humongous code reshuffle.
This does the following:

    * moves octopus::{request,headers,server} into octopus::http
      modules.

    * Moves upstream connection logic into octopus::http::client.

    * Moves Headers and Request serialization to using the
      std::convert::Into trait.

    * Introduces an as-yet-unused http::reply::Reply object.

    * Moves listening server code into a Server struct.
