+++ src/lib.rs
@@ -1,6 +1,4 @@
 #[macro_use]
 pub mod macros;
 
-pub mod headers;
-pub mod request;
-pub mod server;
+pub mod http;
